Beginner level from 1000 To 2000!

+++ src/begginerLvevelSecondHalf/PositiveProducts.java
@@ -9,13 +9,13 @@ public class PositiveProducts {
         int t = sc.nextInt();
         while (t--> 0) {
             int n = sc.nextInt();
-            int arr[] = new int[n];
+            long arr[] = new long[n];
             for (int i = 0; i < n; i++){
                 arr[i] = sc.nextInt();
             }
 
-            int positiveNum = 0;
-            int negativeNum = 0;
+            long positiveNum = 0;
+            long negativeNum = 0;
             for (int i = 0; i < n; i++){
                 if (arr[i] == 0){
                     continue;
